+++ lib/kb_orthofinder/kb_orthofinderImpl.py
@@ -59,13 +59,13 @@ class kb_orthofinder:
                 y=float(y)
                 x=float(x)
 
-                if(group not in Data):
+                if group not in Data:
                     Data[group]=list()
 
                 Data[group].append((x,y))
         fig_gen = GenerateFigure(Data)
 
-        uuid_string = str(uuid.uuid4())
+        uuid_string = "generate_figure_"+str(uuid.uuid4())
         figure_data_file_path=os.path.join(self.scratch,uuid_string)
         os.mkdir(figure_data_file_path)
         fig_gen.generate_figure(figure_path=figure_data_file_path,data_point=params)
@@ -76,21 +76,22 @@ class kb_orthofinder:
         print(('\n' if prefix_newline else '') + time_str + ': ' + message)
 
     def compute_clusters(self, cluster):
+
         features = sorted(cluster.keys())
         clustered_sequences=dict()
         for i in range(len(features)-1):
             for j in range(1,len(features)):
-                if(i>=j):
+                if i>=j:
                     continue
 
                 Spp1=features[i].split("||")[0]
                 Spp2=features[j].split("||")[0]
-                if(Spp1 == features[i]):
+                if Spp1 == features[i]:
                     Spp1 = "_".join(features[i].split("_")[0:-1])
-                if(Spp2 == features[j]):
+                if Spp2 == features[j]:
                     Spp2 = "_".join(features[j].split("_")[0:-1])
 
-                if(Spp1 == Spp2):
+                if Spp1 == Spp2:
                     continue
 
                 Seq1 = cluster[features[i]]
@@ -98,10 +99,10 @@ class kb_orthofinder:
 
                 AA_Match=0.0;
                 for k in range(len(Seq1)):
-                    if(Seq1[k] == "-" or Seq2[k] == "-"):
+                    if Seq1[k] == "-" or Seq2[k] == "-":
                         continue
 
-                    if(Seq1[k] == Seq2[k]):
+                    if Seq1[k] == Seq2[k]:
                         AA_Match+=1.0
 
                 Seq1=Seq1.replace('-','')
@@ -111,12 +112,12 @@ class kb_orthofinder:
                 ID2=AA_Match/len(Seq2)
                 Avg = "{0:.6f}".format((ID1+ID2)/2.0)
 
-                if("Athaliana" in Spp1):
-                    if(features[j] not in clustered_sequences):
+                if "Athaliana" in Spp1:
+                    if features[j] not in clustered_sequences:
                         clustered_sequences[features[j]]=dict()
                     clustered_sequences[features[j]][features[i]]=Avg
-                elif("Athaliana" in Spp2):
-                    if(features[i] not in clustered_sequences):
+                elif "Athaliana" in Spp2:
+                    if features[i] not in clustered_sequences:
                         clustered_sequences[features[i]]=dict()
                     clustered_sequences[features[i]][features[j]]=Avg
 
@@ -127,22 +128,22 @@ class kb_orthofinder:
         top_orthologs = dict()
         for ortholog in (cluster.keys()):
             
-            if(cluster[ortholog] not in top_orthologs):
+            if cluster[ortholog] not in top_orthologs:
                 top_orthologs[cluster[ortholog]]=dict()
             top_orthologs[cluster[ortholog]][ortholog]=1
 
         top_ortholog_seqid="0.00"
         for seq_id in (top_orthologs.keys()):
-            if(float(seq_id)>float(top_ortholog_seqid)):
+            if float(seq_id)>float(top_ortholog_seqid):
                 top_ortholog_seqid=seq_id
 
         #Key to controlling the propagation of annotation
         #Is the sequence identity of the highest-scoring ortholog good enough?
-        if(float(top_ortholog_seqid) < threshold):
+        if float(top_ortholog_seqid) < threshold:
             return (None,"Unannotated 1: LESS_THAN_THRESHOLD",0.0)
 
         top_ortholog=""
-        if(len(top_orthologs[top_ortholog_seqid])==1):
+        if len(top_orthologs[top_ortholog_seqid])==1:
             top_ortholog = list(top_orthologs[top_ortholog_seqid].keys())[0]
         else:
             #There are multiple Arabidopsis orthologs that have the same level of seq. id
@@ -153,7 +154,7 @@ class kb_orthofinder:
             for ortholog in (top_orthologs[top_ortholog_seqid].keys()):
                 function="Unannotated"
                 ortholog_gene = ".".join(ortholog.split('.')[0:-1])
-                if(ortholog_gene not in plantseed_curation):
+                if ortholog_gene not in plantseed_curation:
                     function = "Unannotated 2: ARA_GENE_NOT_IN_CURATION"
                     Unannotated = True
                 else:
@@ -169,33 +170,33 @@ class kb_orthofinder:
             # 3) if 2 functions, and none is "Unannotated" or > 2 functions, 
             #    its ambiguous, return without doing anything
             is_ambiguous=False
-            if(len(Multi_Functions.keys())==1):
+            if len(Multi_Functions.keys())==1:
                 top_ortholog = list(top_orthologs[top_ortholog_seqid].keys())[0]
-            elif(len(Multi_Functions.keys())==2):
-                if(Unannotated is True):
+            elif len(Multi_Functions.keys())==2:
+                if Unannotated is True:
                     for ortholog in (top_orthologs[top_ortholog_seqid].keys()):
                         ortholog_gene = ".".join(ortholog.split('.')[0:-1])
-                        if(ortholog_gene not in plantseed_curation):
+                        if ortholog_gene not in plantseed_curation:
                             continue
                         top_ortholog = ortholog
                         break
                 else:
                     functions_list = sorted(list(Multi_Functions.keys()))
-                    if( functions_list[0] in functions_list[1] or \
-                            functions_list[1] in functions_list[0] ):
+                    if functions_list[0] in functions_list[1] or \
+                            functions_list[1] in functions_list[0]:
                         top_ortholog = list(top_orthologs[top_ortholog_seqid].keys())[0]
                     else:
                         is_ambiguous=True
             else:
                 is_ambiguous=True
 
-            if(is_ambiguous is True):
+            if is_ambiguous is True:
                 #Ambiguously curated top orthologs, so pass
                 print("Ambiguous Functions: ","||".join(list(Multi_Functions.keys())))
                 return (None,"Unannotated 3: AMB_HIT",0.0)
 
         top_ortholog_gene = ".".join(top_ortholog.split('.')[0:-1])
-        if(top_ortholog_gene in plantseed_curation):
+        if top_ortholog_gene in plantseed_curation:
             return (top_ortholog,plantseed_curation[top_ortholog_gene]['function'],top_ortholog_seqid)
         else:
             return (None,"Unannotated 4: NO_ARA_HIT",0.0)
@@ -208,12 +209,12 @@ class kb_orthofinder:
         #BEGIN_CONSTRUCTOR
         self.workspaceURL = config['workspace-url']
 
-        self.testing = False
-        if(config['testing'] == '1'):
-            self.testing=True
+        self.skip_refdata = False
+        if config['skip_refdata'] == '1':
+            self.skip_refdata=True
 
         self.runOrthoFinder = True
-        if(config['run_orthofinder'] == '0'):
+        if config['run_orthofinder'] == '0':
             self.runOrthoFinder=False
             
         self.token = os.environ['KB_AUTH_TOKEN']
@@ -240,20 +241,25 @@ class kb_orthofinder:
 
         output = dict()
 
+        #Create directory for storing temporary files
+        uuid_string = str(uuid.uuid4())
+        uuid_file_path=os.path.join(self.scratch,uuid_string)
+        os.mkdir(uuid_file_path)
+
         # Retrieve plant genome
         self.log("Fetching plant genome: "+input['input_ws']+'/'+input['input_genome'])
         plant_genome = self.dfu.get_objects({'object_refs': [input['input_ws']+'/'+input['input_genome']]})['data'][0]
 
         #Need to extract longest CDS, but only if CDSs available
         use_cds=1
-        if(len(plant_genome['data']['cdss'])==0):
+        if len(plant_genome['data']['cdss'])==0:
             use_cds=0
-            if(len(plant_genome['data']['features'])==0):
+            if len(plant_genome['data']['features'])==0:
                 raise Exception("The genome does not contain any CDSs or features!")
 
         #Now need to be able to retrieve cds entity
         child_cds_index = dict()
-        if(use_cds==1):
+        if use_cds==1:
             child_cds_index = dict([(f['id'], i) for i, f in enumerate(plant_genome['data']['cdss'])])
 
         # If use_cds==1 iterate through features, iterate through CDSs, find longest sequence, use parent mRNA ID    
@@ -261,47 +267,55 @@ class kb_orthofinder:
         self.log("Collecting protein sequences")
         sequences_dict=dict()
         for ftr in plant_genome['data']['features']:
-            if(use_cds==0 and len(ftr['protein_translation'])>0):
+            if use_cds==0 and len(ftr['protein_translation'])>0:
                 sequences_dict[ftr['id']]=ftr['protein_translation']
-            if(use_cds==1 and 'cdss' in ftr):
+            if use_cds==1 and 'cdss' in ftr:
                 longest_sequence=""
                 longest_sequence_id=""
                 for cds_id in ftr['cdss']:
                     sequence = plant_genome['data']['cdss'][child_cds_index[cds_id]]['protein_translation']
-                    if(len(sequence) > len(longest_sequence)):
+                    if len(sequence) > len(longest_sequence):
                         longest_sequence = sequence
                         longest_sequence_id = cds_id
-                if(len(longest_sequence)>0):
+                if len(longest_sequence)>0:
                     sequences_dict[longest_sequence_id]=longest_sequence
 
-        if(len(sequences_dict)==0):
+        if len(sequences_dict)==0:
             raise Exception("The genome does not contain any protein sequences!")
 
-        output['ftrs'] = len(sequences_dict.keys())
+        # We have a problem with how OrthoFinder arbitrarily replaces 'special' characters in identifiers:
+        # https://github.com/davidemms/OrthoFinder/blob/master/scripts_of/util.py#L181
+        # without 'hacking' the OF code, here I'm trying to anticipate how OF may have changed the identifier
 
-        #Create directory for storing new fasta file
-        uuid_string = str(uuid.uuid4())
-        fasta_file_path=os.path.join(self.scratch,uuid_string)
-        os.mkdir(fasta_file_path)
+        alt_seq_ids_dict=dict()
+        for og_id in sequences_dict:
+            of_id = og_id.replace(":", "_").replace(",", "_").replace("(", "_").replace(")", "_")
+            alt_seq_ids_dict[of_id]=og_id
+
+        output['ftrs'] = len(sequences_dict.keys())
 
         #Reference data is considered immutable but each run modifies results within the directory
         #So here, we copy the reference data directory into scratch
         #The first if condition is for testing purposes
+        family_file_path=os.path.join(uuid_file_path,"family_data")
 
-        uuid_string = str(uuid.uuid4())
-        family_file_path=os.path.join(self.scratch,uuid_string,"Reference_Results")
-
-        if(self.testing is True):
-            if('families_path' in input and os.path.isdir(input['families_path'])):
-                family_file_path = input['families_path']
-            self.log("Testing Reference Families at "+family_file_path)
+        if self.skip_refdata:
+            if 'families_path' in input and os.path.isdir(input['families_path']):
+                self.log("Copying Test Families at "+family_file_path)
+                shutil.copytree(input['families_path'],family_file_path)
         else:
             self.log("Copying Reference Families to "+family_file_path)
             shutil.copytree("/data/OrthoFinder_Phytozome_Reference",family_file_path)
 
+        # Create directory for storing new fasta file
+        # This has to be a fresh directory for OrthoFinder
+        # to find and integrate the fasta file
+        fasta_file_path=os.path.join(uuid_file_path,"protein_fasta")
+        os.mkdir(fasta_file_path)
+
         #The fasta file must have a random name to avoid _any_ clashes
         #This will need to be replaced in the newick file
-        temp_genome_name = str(uuid.uuid4())
+        temp_genome_name = "protein_fasta_"+str(uuid.uuid4())
         protein_fasta_file = os.path.join(fasta_file_path,temp_genome_name+".fa")
         self.log("Printing protein sequences to file: "+protein_fasta_file)
 
@@ -310,16 +324,16 @@ class kb_orthofinder:
             #Code plagarized from https://github.com/biopython/biopython/blob/master/Bio/SeqIO/FastaIO.py
             for seq_id in sequences_dict:
                 #printing smaller set for testing purposes
-                if(self.testing is True):
+                if self.skip_refdata:
                     testing_count = testing_count -1
                 fasta_handle.write(">"+seq_id+"\n")
                 for i in range(0, len(sequences_dict[seq_id]), 80):
                     fasta_handle.write(sequences_dict[seq_id][i:i+80]+"\n")
-                if(testing_count==0):
+                if testing_count==0:
                     break
                 
         #Building command
-        command = "/kb/deployment/bin/orthofinder/orthofinder.py "
+        command = "/usr/bin/orthofinder/orthofinder.py "
         #Software
         command +="-S diamond -M msa -A mafft -T fasttree "
         # No. of Threads
@@ -335,7 +349,7 @@ class kb_orthofinder:
 
         #####################################################
         output_files=list()
-        if(self.testing is False or self.runOrthoFinder is True):
+        if self.skip_refdata is False or self.runOrthoFinder is True:
             self.log("Running OrthoFinder command: "+command)
 
             pipe = subprocess.Popen(command, stdout=subprocess.PIPE, shell=True)
@@ -379,11 +393,11 @@ class kb_orthofinder:
         output['fns']=list()
         for feature in plantseed_curation:
             function = plantseed_curation[feature]['function']
-            if( function not in output['fns'] ):
+            if  function not in output['fns']:
                 output['fns'].append(function)
 
             for role in plantseed_curation[feature]['roles']:
-                if(role not in Ignored_Curation):
+                if role not in Ignored_Curation:
                     PlantSEED_Roles[role]=1
 
         output['fns']=len(output['fns'])
@@ -416,20 +430,24 @@ class kb_orthofinder:
                     transcript_id = header
                     transcript_description = None
 
+                # anticipating if OF changed characters in ids, see above
+                if transcript_id in alt_seq_ids_dict:
+                    transcript_id = alt_seq_ids_dict[transcript_id]
+
                 #skip un-necessary proteins to reduce computation time
-                if("Athaliana" not in transcript_id and transcript_id not in sequences_dict):
+                if "Athaliana" not in transcript_id and transcript_id not in sequences_dict:
                     continue
 
                 seq = seq.upper()
                 family_sequences_dict[transcript_id]=seq
                 gene_id = ".".join(transcript_id.split('.')[0:-1])
-                if(gene_id in plantseed_curation):
+                if gene_id in plantseed_curation:
                     function = plantseed_curation[gene_id]['function']
                     curated_family.append(function+"|||"+transcript_id)
                     output['alignments'].append(transcript_id)
             
-            if(len(curated_family)>0):
-                if(family not in families_dict):
+            if len(curated_family)>0:
+                if family not in families_dict:
                     families_dict[family]=dict()
                 families_dict[family]['sequences']=family_sequences_dict
                 families_dict[family]['functions']=curated_family
@@ -451,14 +469,14 @@ class kb_orthofinder:
 
             for function_ortholog in families_dict[family]['functions']:
                 (function,ortholog) = function_ortholog.split("|||")
-                if(function not in functions_dict):
+                if function not in functions_dict:
                     functions_dict[function]=dict()
-                if(family not in functions_dict[function]):
+                if family not in functions_dict[function]:
                     functions_dict[function][family]={'orthologs':[],
                                                       'hits':[],
                                                       'cluster':pw_seq_id_list}
 
-                if(ortholog not in functions_dict[function][family]['orthologs']):
+                if ortholog not in functions_dict[function][family]['orthologs']:
                     functions_dict[function][family]['orthologs'].append(ortholog)
 
             for spp_ftr in sorted(pw_seq_id_list.keys()):
@@ -468,31 +486,31 @@ class kb_orthofinder:
                 ftr = spp_ftr.replace(temp_genome_name+"_","")
 
                 # Write out results
-                if(function not in functions_dict):
+                if function not in functions_dict:
                     annotate_fh.write("MISSING FUNCTION: "+" | ".join([function,family,str(ortholog),spp_ftr])+"\n")
                 else:
                     annotate_fh.write("FOUND FUNCTION: "+" | ".join([function,family,str(ortholog),spp_ftr])+"\n")
-                    if(family not in functions_dict[function]):
+                    if family not in functions_dict[function]:
                         annotate_fh.write("MISSING FAMILY: "+" | ".join([function,family,str(ortholog),spp_ftr])+"\n")
                     else:
                         annotate_fh.write("FOUND FAMILY: "+" | ".join([function,family,str(ortholog),spp_ftr])+"\n")
-                        if(ortholog not in functions_dict[function][family]['orthologs']):
+                        if ortholog not in functions_dict[function][family]['orthologs']:
                             annotate_fh.write("MISSING ORTHOLOG: "+" | ".join([function,family,str(ortholog),spp_ftr])+"\n")
                         else:
                             annotate_fh.write("FOUND ORTHOLOG: "+" | ".join([function,family,str(ortholog),spp_ftr])+"\n")
 
                 # Save result
-                if("Unannotated" in function):
+                if "Unannotated" in function:
                     function="Unannotated"
 
                 annotated_features_dict[ftr]=function
 
-                if(function not in found_annotations and "Unannotated" not in function):
+                if function not in found_annotations and "Unannotated" not in function:
                     found_annotations.append(function)
 
-                if(function in functions_dict and \
+                if function in functions_dict and \
                        family in functions_dict[function] and \
-                       ortholog in functions_dict[function][family]['orthologs']):
+                       ortholog in functions_dict[function][family]['orthologs']:
                     functions_dict[function][family]['hits'].append({'seqid':seqid,
                                                                      'feature':spp_ftr,
                                                                      'ortholog':ortholog})
@@ -511,7 +529,7 @@ class kb_orthofinder:
         #But, if annotating CDS, need to be able to retrieve parent feature/transcripts
         parent_feature_index = dict()
         parent_transcript_index = dict()
-        if(use_cds==1):
+        if use_cds==1:
             parent_feature_index = dict([(f['id'], i) for i, f in enumerate(plant_genome['data']['features'])])
             parent_transcript_index = dict([(f['id'], i) for i, f in enumerate(plant_genome['data']['mrnas'])])
 
@@ -520,11 +538,11 @@ class kb_orthofinder:
         # As the Phytozome genomes have CDSs, the features don't usually get annotated here
         for ftr in plant_genome['data']['features']:
             ftr['functions']=["Unannotated"]
-            if(ftr['id'] in annotated_features_dict):
+            if ftr['id'] in annotated_features_dict:
                 ftr['functions']=[annotated_features_dict[ftr['id']]]
 
             # It is possible that a gene is listed without an associated transcript
-            if('mrnas' in ftr):
+            if 'mrnas' in ftr:
                 
                 # Add annotation to transcripts
                 # As the Phytozome genomes have CDSs, the features don't usually get annotated here
@@ -538,14 +556,14 @@ class kb_orthofinder:
                     mrna_obj['functions'] = [ftr['functions'][0]]
 
                     # If it happens that the mRNA is independently annotated
-                    if(mrna in annotated_features_dict):
+                    if mrna in annotated_features_dict:
                         mrna_obj['functions'] = [annotated_features_dict[mrna]]
 
                         # Then annotate parent feature gene
                         ftr['functions']=[annotated_features_dict[mrna]]
                         
             # It is possible that a gene is listed without an associated protein
-            if('cdss' in ftr):
+            if 'cdss' in ftr:
 
                 # Add annotation to proteins
                 # As the Phytozome genomes have CDSs, the features and mrnas get annotated here
@@ -560,10 +578,10 @@ class kb_orthofinder:
 
                     # If it happens that the CDS is independently annotated
                     # Which is most likely event if using Phytozome genomes
-                    if(cds in annotated_features_dict):
+                    if cds in annotated_features_dict:
                         cds_obj['functions'] = [annotated_features_dict[cds]]
 
-                        if('parent_mrna' in cds_obj):
+                        if 'parent_mrna' in cds_obj:
                             parent_transcript_indice = parent_transcript_index[cds_obj['parent_mrna']]
                             parent_transcript_obj = plant_genome['data']['mrnas'][parent_transcript_indice]
                             parent_transcript_obj['functions'] = [annotated_features_dict[cds]]
@@ -578,15 +596,17 @@ class kb_orthofinder:
             json.dump(plant_genome, fh)
             fh.close()
 
-        if('output_genome' not in input):
+        if 'output_genome' not in input:
             input['output_genome']=input['input_genome']
 
         saved_genome=""
-        if(self.testing is True):
-            #wsid = self.dfu.ws_name_to_id(input['input_ws'])
-            #save_result = self.dfu.save_objects({'id':wsid,'objects':[{'name':input['output_genome'],
-            #                                                           'data':plant_genome['data'],
-            #                                                           'type':'KBaseGenomes.Genome'}]})[0]
+        if self.skip_refdata is True:
+            # I keep this commented code here just in case I need to debug 
+            # how a genome is saved when testing without refdata
+            # wsid = self.dfu.ws_name_to_id(input['input_ws'])
+            # save_result = self.dfu.save_objects({'id':wsid,'objects':[{'name':input['output_genome'],
+            #                                                            'data':plant_genome['data'],
+            #                                                            'type':'KBaseGenomes.Genome'}]})[0]
             pass
         else:
             save_result = self.gfu.save_one_genome({'workspace' : input['input_ws'],
@@ -605,7 +625,7 @@ class kb_orthofinder:
             Function = Function_Comments.pop(0)
             Roles = re.split("\s*;\s+|\s+[\@\/]\s+", Function)
             for role in Roles:
-                if(role in PlantSEED_Roles):
+                if role in PlantSEED_Roles:
                     Annotated_Roles[role]=1
 
         output['hit_fns']=len(found_annotations)
@@ -616,8 +636,7 @@ class kb_orthofinder:
         fraction_plantseed = float( (float(len(Annotated_Roles.keys())) / float(len(PlantSEED_Roles.keys()))) )
 
         # HTML Folder Path
-        uuid_string = str(uuid.uuid4())
-        html_file_path=os.path.join(self.scratch,uuid_string)
+        html_file_path=os.path.join(uuid_file_path,"generate_report")
         os.mkdir(html_file_path)
 
         # Generate figure: 
@@ -726,13 +745,13 @@ class kb_orthofinder:
                           'file_links' : output_files,
                           'html_links' : html_report_list}
 
-        if(self.testing is False):
+        if self.skip_refdata is False:
             report_params['objects_created']=[{"ref":saved_genome,"description":description}]
 
-        kbase_report_client = KBaseReport(self.callback_url, token=self.token)
-        report_client_output = kbase_report_client.create_extended_report(report_params)
-        output['report_name']=report_client_output['name']
-        output['report_ref']=report_client_output['ref']
+            kbase_report_client = KBaseReport(self.callback_url, token=self.token)
+            report_client_output = kbase_report_client.create_extended_report(report_params)
+            output['report_name']=report_client_output['name']
+            output['report_ref']=report_client_output['ref']
 
         #END annotate_plant_transcripts